requirements.txt
streamlit>=1.0
python-gnupg>=0.4.9
